% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/all_metrics.R
\name{all_metrics}
\alias{all_metrics}
\title{Calculate All Metrics}
\usage{
all_metrics(Long, isd.df)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{isd.df}{= A data frame containing all of the Impact Source
Determination (ISD) models.}
}
\value{
Calculates all of the metrics in the 2014 NYSDEC SOP.
}
\description{
Calculate All Metrics
}

