% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_jab.R
\name{bap_jab}
\alias{bap_jab}
\title{BAP Jab}
\usage{
bap_jab(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for jabs.
}
\description{
BAP Jab
}

