% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_low_gradient.R
\name{bap_low_gradient}
\alias{bap_low_gradient}
\title{BAP Low Gradient}
\usage{
bap_low_gradient(Long)
}
\arguments{
\item{Long}{Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for the Low Gradient IBI.
}
\description{
BAP Low Gradient
}
