% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_mp_nav_water.R
\name{bap_mp_nav_waters}
\alias{bap_mp_nav_waters}
\title{BAP Multiple-Plate Samples Navigable Waters}
\usage{
bap_mp_nav_waters(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for
multiple-plate samples collected in navigable waters.
}
\description{
BAP Multiple-Plate Samples Navigable Waters
}
