% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_mp_non_nav_water.R
\name{bap_mp_non_nav_water}
\alias{bap_mp_non_nav_water}
\title{BAP Multiple-Plate Samples Non-Navigable Waters}
\usage{
bap_mp_non_nav_water(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for
multiple-plate samples collected in non-navigable waters.
}
\description{
BAP Multiple-Plate Samples Non-Navigable Waters
}
