% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_ponar.R
\name{bap_ponar}
\alias{bap_ponar}
\title{BAP Ponar}
\usage{
bap_ponar(long)
}
\arguments{
\item{long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for ponars.
}
\description{
BAP Ponar
}
