% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bap_riffle.R
\name{bap_riffle}
\alias{bap_riffle}
\title{BAP Riffle}
\usage{
bap_riffle(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate and score the metrics in the BAP created for riffles.
}
\description{
BAP Riffle
}
