% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd.R
\name{calc.isd}
\alias{calc.isd}
\title{Calculate All of the ISD Models}
\usage{
calc.isd(Long, isd.df)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{isd.df}{= a data frame with the ISD models.}
}
\value{
The percentage of similarity the observed community has
to each of the model communities.
}
\description{
Calculate All of the ISD Models
}

