% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_metrics.R
\name{coe_rich}
\alias{coe_rich}
\title{COE Richness}
\usage{
coe_rich(Long, Level)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Level}{= Must specify either 'FAMILY' or "GENUS.'}
}
\value{
The number of Coleoptera, Odonata, and Ephmeroptera (COE)
}
\description{
COE Richness
}
