% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{data_prep}
\alias{data_prep}
\title{Prepare the Data}
\usage{
data_prep(Taxa.df)
}
\arguments{
\item{Taxa.df}{= A data frame containing raw taxonomic counts.}
}
\value{
Prepare taxonomic data for metric calculations.
}
\description{
Prepare the Data
}

