% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_metrics.R
\name{ept_rich}
\alias{ept_rich}
\title{EPT Richness Excluding Tolerant Taxa}
\usage{
ept_rich(Long, Level)
}
\arguments{
\item{Long}{Taxonomic count data in a long data format.}

\item{Level}{Typically, "FINAL_ID".}
}
\value{
The number of Ephemeroptera, Plecoptera, and Trichoptera (EPT)
taxa excluding tolerant taxa.
}
\description{
EPT Richness Excluding Tolerant Taxa
}
