% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{event_prep}
\alias{event_prep}
\title{Create an Event_ID}
\usage{
event_prep(data.df)
}
\arguments{
\item{Check.df}{=}
}
\value{
Create a unique event identification (EVENT_ID).
}
\description{
Create an Event_ID
}
