% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{group_rich}
\alias{group_rich}
\title{Vector of taxa richness for a specific list of taxa}
\usage{
group_rich(NameList, taxa.df)
}
\arguments{
\item{NameList}{= uninque list of taxa.}

\item{taxa.df}{= Wide data frame format of taxonomic counts.}
}
\value{
A vector of taxa richness for a specific list of taxa representing
each sampling event. NameList from the wide data frame of taxa
(i.e. Family level or Genus level)
}
\description{
Vector of taxa richness for a specific list of taxa
}

