% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_metrics.R
\name{insect_rich}
\alias{insect_rich}
\title{Insect Richness}
\usage{
insect_rich(Long, Level)
}
\arguments{
\item{Long}{Taxonomic count data in a long data format.}

\item{Level}{Typically, "FINAL_ID".}
}
\value{
The number of insect (INSECTA) taxa.
}
\description{
Insect Richness
}
