% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd.R
\name{isd.models}
\alias{isd.models}
\title{ISD Models}
\usage{
isd.models(isd.df, comm.df)
}
\arguments{
\item{isd.df}{= a data frame with the ISD models.}

\item{comm.df}{= a data frame of observed community percentages.}
}
\value{
The percentage of similarity the observed community has
to each model community.
}
\description{
ISD Models
}

