% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isd.R
\name{isd_prep}
\alias{isd_prep}
\title{ISD Preparation}
\usage{
isd_prep(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
A data frame of the percentage each taxon, necessary for ISD
calculations, in each sampling event.
}
\description{
ISD Preparation
}

