% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/acid_tol_index.R
\name{pct_acid_tol}
\alias{pct_acid_tol}
\title{Acid Tolerance Index (ATI)}
\usage{
pct_acid_tol(Genus)
}
\arguments{
\item{Genus}{= Genus level taxa count data.}
}
\value{
The percentage of Acid Tolerant Individuals (ATI).
}
\description{
Acid Tolerance Index (ATI)
}
