% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_dom3.R
\name{pct_dom3}
\alias{pct_dom3}
\title{Percentage of the 3rd Dominant Family}
\usage{
pct_dom3(Taxa.df)
}
\arguments{
\item{Taxa.df}{= Wide data frame format of taxonomic counts.}
}
\value{
Percent of individuals that represent the first, second, and third
most abundant taxa.
}
\description{
Percentage of the 3rd Dominant Family
}

