% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_rich_ephem.R
\name{pct_rich_ephem_epeorus}
\alias{pct_rich_ephem_epeorus}
\title{Percent Ephemeroptera Richness minus Epeorus}
\usage{
pct_rich_ephem_epeorus(Long, Genus)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Genus}{= Genus level taxa count data}
}
\value{
The percentage of the taxa observed  represented by Ephemeroptera taxa
minus the genus Epeorus.
}
\description{
Percent Ephemeroptera Richness minus Epeorus
}
