% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma.R
\name{pct_taxon}
\alias{pct_taxon}
\title{Percentage of a Taxon}
\usage{
pct_taxon(Taxon, Level, long)
}
\arguments{
\item{Taxon}{= The taxon of interest in quotations and in all caps
(e.g. "EPEORUS")}

\item{Level}{= The taxonomic level (rank) the taxon of intrest can be found.}

\item{long}{= Taxonomic count data in a long data format.}
}
\value{
Calculate the percentage of a particular taxon.
}
\description{
Percentage of a Taxon
}
