% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/percent_composition_metrics.R
\name{percent_composition}
\alias{percent_composition}
\title{Calculate the percent composition}
\usage{
percent_composition(Long, Level, target_taxa)
}
\arguments{
\item{Long}{Taxonomic count data in a long data format.}

\item{Level}{The taxonomic level/group our target_taxa can be found in. This should be the name of a column.}

\item{target_taxa}{The taxa we are interested in summarizing.}
}
\description{
Calculate the percent composition
}
