% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma.R
\name{pma}
\alias{pma}
\title{Percent Model Affinity (PMA)}
\usage{
pma(Long)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}
}
\value{
The similarity of the observed community to a model community.
}
\description{
Percent Model Affinity (PMA)
}

