% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma.R
\name{pma_ponar}
\alias{pma_ponar}
\title{Ponar Percent Model Affinity (PMA)}
\usage{
pma_ponar(long, pma.df = pma.ponar)
}
\arguments{
\item{long}{= Taxonomic count data in a long data format.}

\item{pma.df}{= The PMA table.}
}
\value{
The similarity of the observed community to a model community.
}
\description{
Ponar Percent Model Affinity (PMA)
}

