% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prep_master_taxa}
\alias{prep_master_taxa}
\title{Prepare Master Taxa List}
\usage{
prep_master_taxa(master)
}
\arguments{
\item{taxa.df}{= A data frame containing raw taxonomic counts.}
}
\value{
Prepare taxonomic data for metric calculations.
}
\description{
Prepare Master Taxa List
}

