% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pct_rich_ephem.R
\name{rich_ephemeroptera}
\alias{rich_ephemeroptera}
\title{Ephemeroptera Richness}
\usage{
rich_ephemeroptera(Long, Level = "GENUS")
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Level}{= The taxonomic level to preform the analysis.  This metric must
specify a taxonomic rank below order.}
}
\value{
The number of Ephemeroptera taxa.
}
\description{
Ephemeroptera Richness
}
