% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nco_rich.R
\name{rich_nco}
\alias{rich_nco}
\title{Non-Chironomidae and Oligochaeta Taxa Richness}
\usage{
rich_nco(Long, Level = "GENUS")
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Level}{= The taxonomic level to preform the analysis.  This metric must
specify a taxonomic rank below family.}
}
\value{
Taxa richness excluding Chironomidae or Oligochaeta taxa.
}
\description{
Non-Chironomidae and Oligochaeta Taxa Richness
}
