% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/richness_metrics.R
\name{richness}
\alias{richness}
\title{Calculate Richness}
\usage{
richness(Long, Level)
}
\arguments{
\item{Long}{a long data frame.}

\item{Level}{the taxonomic level of interest.}
}
\value{
a vector of richness values.
}
\description{
Calculate Richness
}
