% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jab_metrics.R
\name{score_ept_jab}
\alias{score_ept_jab}
\title{Score EPT Richness (Jab)}
\usage{
score_ept_jab(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw EPT richness calculations with the ranges assigned to
jab samples.
}
\description{
Score EPT Richness (Jab)
}
