% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mp_nav_water_metrics.R
\name{score_ept_mp_nav_water}
\alias{score_ept_mp_nav_water}
\title{Score EPT Richness (Multiple-Plate Navigable Waters)}
\usage{
score_ept_mp_nav_water(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw EPT richness calculations with the ranges assigned to
multiple-plate samples collected in navigable waters.
}
\description{
Score EPT Richness (Multiple-Plate Navigable Waters)
}

