% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_riffle_metrics.R
\name{score_ept_riffle}
\alias{score_ept_riffle}
\title{Score EPT Richness (Riffle)}
\usage{
score_ept_riffle(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw EPT richness calculations with the ranges assigned to
riffle samples.
}
\description{
Score EPT Richness (Riffle)
}
