% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jab_metrics.R
\name{score_hbi_jab}
\alias{score_hbi_jab}
\title{Score HBI (Jab)}
\usage{
score_hbi_jab(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw HBI calculations with the ranges assigned to
jab samples.
}
\description{
Score HBI (Jab)
}
