% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_riffle_metrics.R
\name{score_hbi_riffle}
\alias{score_hbi_riffle}
\title{Score HBI (Riffle)}
\usage{
score_hbi_riffle(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw HBI calculations with the ranges assigned to
riffle samples.
}
\description{
Score HBI (Riffle)
}
