% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_low_gradient.R
\name{score_low_gradient}
\alias{score_low_gradient}
\title{Score Low Gradient Metrics}
\usage{
score_low_gradient(metric, values)
}
\arguments{
\item{metric}{The name of the metric to be scored.}

\item{values}{The raw metric values that correspond with the specified metric arguement.}
}
\description{
Score Low Gradient Metrics
}
