% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_riffle_metrics.R
\name{score_nbip_riffle}
\alias{score_nbip_riffle}
\title{Score NBI-P (Riffle)}
\usage{
score_nbip_riffle(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw NBI-P calculations with the ranges assigned to
riffle samples.
}
\description{
Score NBI-P (Riffle)
}
