% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jab_metrics.R
\name{score_nco_jab}
\alias{score_nco_jab}
\title{Score NCO Richness (Jab)}
\usage{
score_nco_jab(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw NCO Richness calculations with the ranges assigned to
jab samples.
}
\description{
Score NCO Richness (Jab)
}

