% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ponar_metrics.R
\name{score_pct_dom3_ponar}
\alias{score_pct_dom3_ponar}
\title{Score Percent DOM-3 (Ponar)}
\usage{
score_pct_dom3_ponar(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw percent DOM-3 calculations with the ranges assigned to
ponar samples.
}
\description{
Score Percent DOM-3 (Ponar)
}

