% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ponar_metrics.R
\name{score_pma_ponar}
\alias{score_pma_ponar}
\title{Score PMA (Ponar)}
\usage{
score_pma_ponar(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw PMA calculations with the ranges assigned to
ponar samples.
}
\description{
Score PMA (Ponar)
}

