% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_riffle_metrics.R
\name{score_pma_riffle}
\alias{score_pma_riffle}
\title{Score PMA (Riffle)}
\usage{
score_pma_riffle(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw PMA calculations with the ranges assigned to
riffle samples.
}
\description{
Score PMA (Riffle)
}

