% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_jab_metrics.R
\name{score_rich_jab}
\alias{score_rich_jab}
\title{Score Richness (Jab)}
\usage{
score_rich_jab(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw richness calculations with the ranges assigned to
jab samples.
}
\description{
Score Richness (Jab)
}
