% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ponar_metrics.R
\name{score_rich_ponar}
\alias{score_rich_ponar}
\title{Score Richness (Ponar)}
\usage{
score_rich_ponar(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw richness calculations with the ranges assigned to
ponar samples.
}
\description{
Score Richness (Ponar)
}
