% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_riffle_metrics.R
\name{score_rich_riffle}
\alias{score_rich_riffle}
\title{Score Richness (Riffle)}
\usage{
score_rich_riffle(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw richness calculations with the ranges assigned to
riffle samples.
}
\description{
Score Richness (Riffle)
}

