% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_mp_nav_water_metrics.R
\name{score_shannon_mp_nav_water}
\alias{score_shannon_mp_nav_water}
\title{Score Shannon-Wiener Diversity (Multiple-Plate Navigable Waters)}
\usage{
score_shannon_mp_nav_water(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw Shannon-Wiener Diversity calculations with the ranges assigned to
multiple-plate samples collected in navigable waters.
}
\description{
Score Shannon-Wiener Diversity (Multiple-Plate Navigable Waters)
}
