% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_ponar_metrics.R
\name{score_shannon_ponar}
\alias{score_shannon_ponar}
\title{Score Shannon-Wiener Diversity (Ponar)}
\usage{
score_shannon_ponar(metrics.df)
}
\arguments{
\item{metrics.df}{= a data frame of calculated metrics values for each
sampling event.}
}
\value{
Scores the raw Shannon-Wiener Diversity calculations
with the ranges assigned to ponar samples.
}
\description{
Score Shannon-Wiener Diversity (Ponar)
}

