% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{taxa_prep3}
\alias{taxa_prep3}
\title{Prepare the Taxonomic Count Data}
\usage{
taxa_prep3(taxa.df)
}
\arguments{
\item{taxa.df}{= A data frame containing raw taxonomic counts.}
}
\value{
Prepare taxonomic data for metric calculations.
}
\description{
Prepare the Taxonomic Count Data
}

