% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tol_indices.R
\name{tol_index}
\alias{tol_index}
\title{Tolerance Indices (Hilsenhoff, NBI_P, NBI_N)}
\usage{
tol_index(Long, Index = "TOLERANCE", Level = "SAMPLES_GENUS_SPECIES")
}
\arguments{
\item{Long}{= Taxonomic count data in a Long data format.}

\item{Index}{= The column with the tolerance values of interest.}

\item{Level}{= Taxonomic rank used for calculations.}
}
\value{
The average tolerance score per taxon for each unique sampling event.
If taxon does not have an assigned tolerance value it is excluded from the analysis.
}
\description{
Tolerance Indices (Hilsenhoff, NBI_P, NBI_N)
}
