% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{wide}
\alias{wide}
\title{Wide Data Frame
Transform taxa count data from long to wide format}
\usage{
wide(Long, Level)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Level}{= Taxonomic Level (PHYLUM, CLASS, ORDER, FAMILY, GENUS).}
}
\value{
Taxa counts in wide data format.
}
\description{
Wide Data Frame
Transform taxa count data from long to wide format
}

