% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ept_rich.R
\name{ept_rich}
\alias{ept_rich}
\title{EPT Richness Excluding Tolerant Taxa}
\usage{
ept_rich(Long, Level)
}
\arguments{
\item{Long}{= Taxonomic count data in a long data format.}

\item{Level}{= Must sepecify either 'FAMILY' or "GENUS.'}
}
\value{
The number of Ephemeroptera, Plecoptera, and Trichoptera (EPT)
taxa excluding tolerant taxa.
}
\description{
EPT Richness Excluding Tolerant Taxa
}
