% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pma.R
\name{pma}
\alias{pma}
\title{Percent Model Affinity (PMA)}
\usage{
pma(long, pma.df = pma.model)
}
\arguments{
\item{long}{= Taxonomic count data in a long data format.}

\item{pma.df}{= The PMA table.}
}
\value{
The similarity of the observed community to a model community.
}
\description{
Percent Model Affinity (PMA)
}
